Mock more than one dependency
Mock both the instance class and the static class in a test case.
This is actually the first time, in this repo, that I mock an instance
object via specifying a return value for the constructor (...I think).
I'm not up-to-speed on the details behind the curtains so I don't really
understand what's going on. Maybe there's a nicer way of doing it. The
current mock setup of the instance class pretty much follows the example
at the top of the "Quick Guide" in the Python documentation's chapter on
mocking.